[C#] refactor: move handlers to separate files for readability
- move adaptive card handlers and message extension handlers to separate
  files
- add type constraint for TState in handlers
